Add custom hook and list component after code review
