+++ src/App.js
@@ -1,43 +1,14 @@
-import { useState } from 'react'
-import { v4 as uuid } from 'uuid'
+import useList from './hooks/useList'
 
-import ListItem from './01.UI-Elements/ListItem'
+import List from './02.Components/List'
 import InputForm from './01.UI-Elements/InputForm'
 
 export default function App() {
-  const [list, setList] = useState({
-    byId: {},
-    allIds: [],
-  })
-
+  const { list, addListItem } = useList()
   return (
     <div>
-      {list.allIds.map((id) => {
-        const { titleListItem, isCheckmarked } = list.byId[id]
-        return (
-          <ListItem
-            key={id}
-            titleListItem={titleListItem}
-            isCheckmarked={isCheckmarked}
-          />
-        )
-      })}
+      <List list={list} />
       <InputForm addListItem={addListItem} />
     </div>
   )
-
-  function addListItem(titleListItem) {
-    const generatedId = uuid()
-    setList({
-      byId: {
-        ...list.byId,
-        [generatedId]: {
-          id: generatedId,
-          titleListItem,
-          isCheckmarked: false,
-        },
-      },
-      allIds: [...list.allIds, generatedId],
-    })
-  }
 }
